feat: create a new file for offset functions
