+++ demo/index.js
@@ -1,11 +1,9 @@
+import { applyOffset, processGeometry } from "@/index"
 import "./index.css"
 import * as THREE from "three"
 
 import { TrackballControls } from "three/examples/jsm/controls/TrackballControls.js"
 import { STLLoader } from "three/examples/jsm/loaders/STLLoader.js"
-import { STLExporter } from "three/examples/jsm/exporters/STLExporter.js"
-import { createOffsetMesh } from "../src/utils/offsetObjectHash"
-import { createMeshFromObject } from "../src/utils/createMeshFromObject"
 
 let renderer, scene
 
@@ -92,7 +90,10 @@ btnMeshOffset.addEventListener("click", async (e) => {
 
   const offset = offsetElement.value
 
-  if (mesh) await applyOffset(mesh, offset)
+  if (mesh) {
+    const meshOffset = await applyOffset(mesh, offset)
+    removeAddOffset(meshOffset)
+  }
 })
 
 btnPointOffset.addEventListener("click", (e) => {
@@ -100,7 +101,10 @@ btnPointOffset.addEventListener("click", (e) => {
 
   const offset = offsetElement.value
 
-  if (mesh) processGeometry(mesh.geometry, offset)
+  if (mesh) {
+    const newMesh = processGeometry(mesh.geometry, offset)
+    removeAddOffset(newMesh)
+  }
 })
 
 // Load the file and get the geometry
@@ -129,58 +133,6 @@ const createMeshFromFile = (geometry) => {
   mesh = new THREE.Mesh(geometry, material)
 
   scene.add(mesh)
-}
-
-/**
- * Creates a mesh with the offset passed
- * @param {THREE.Mesh} meshToOffset The mesh that the user added to the scene
- * @param {Number} offset The offset passed
- */
-const applyOffset = async (meshToOffset, offset) => {
-  console.time()
-
-  // 1. Export mesh as an ascii file
-  const exporter = new STLExporter()
-  const result = exporter.parse(meshToOffset, { binary: false })
-
-  // 2. Creates mesh with offset
-  const object = createOffsetMesh(result, offset)
-  const meshOffset = await createMeshFromObject(object)
-
-  meshOffset.name = "offset"
-
-  console.timeEnd()
-
-  removeAddOffset(meshOffset)
-}
-
-/**
- * Creates a points with the offset passed
- * @param {THREE.Geometry} geometry The geometry that the user added to the scene
- * @param {Number} offset The offset passed
- */
-const processGeometry = (geometry, offset) => {
-  const vertices = geometry.attributes.position.array
-
-  const normals = geometry.attributes.normal.array
-
-  const position = new Float32Array(vertices.length * 3)
-
-  for (let i = 0; i < vertices.length; i = i + 3) {
-    position[i] = offset * normals[i] + vertices[i]
-    position[i + 1] = offset * normals[i + 1] + vertices[i + 1]
-    position[i + 2] = offset * normals[i + 2] + vertices[i + 2]
-  }
-
-  const newGeometry = new THREE.BufferGeometry()
-  newGeometry.setAttribute("position", new THREE.BufferAttribute(position, 3))
-
-  const newMaterial = new THREE.PointsMaterial({ color: 0xff0000, size: 0.7 })
-  const newMesh = new THREE.Points(newGeometry, newMaterial)
-
-  newMesh.name = "offset"
-
-  removeAddOffset(newMesh)
 }
 
 const removeAddOffset = (mesh) => {
