+++ src/index.js
@@ -1,0 +1,3 @@
+export { applyOffset, processGeometry } from "./offset.js"
+export { createMeshFromObject } from "./utils/createMeshFromObject.js"
+export { createOffsetMesh } from "./utils/offsetObjectHash.js"
